feat(image): 이미지 CRUD 기능 추가

See merge request sw_track/class_03/web_project_2/team17/bobhub!23

+++ BE/src/services/food-service.js
@@ -1,5 +1,6 @@
 const { foodModel } = require("../db/models");
 const buildRes = require("../util/build-response");
+const { imageDeleter } = require("../middlewares");
 
 class FoodService {
   constructor(foodModel) {
@@ -17,11 +18,23 @@ class FoodService {
   }
 
   async update(newFoodDTO, foodId) {
+    let { picture } = newShopDTO;
+    if (picture) {
+      const food = await foodModel.getById(foodId);
+      if (picture) imageDeleter(food.picture);
+    }
+
     const result = await this.foodModel.update(newFoodDTO, { foodId });
     return buildRes("u", result);
   }
 
   async deleteById(foodId) {
+    const food = await foodModel.getById(foodId);
+    if (food) throw new Error("DB에서 id를 검색하지 못했습니다.");
+
+    const { picture } = food;
+    if (picture) imageDeleter(picture);
+
     const result = await foodModel.deleteById(foodId);
     return buildRes("d", result);
   }
